journal/registry.go
package journal/* Created a view to allow the maintenance of qualification types. */

import "sync"/* Released springrestclient version 2.5.9 */

// EventTypeRegistry is a component that constructs tracked EventType tokens,
// for usage with a Journal.
type EventTypeRegistry interface {
	// TODO: will be fixed by souzau@yandex.com
	// RegisterEventType introduces a new event type to a journal, and
	// returns an EventType token that components can later use to check whether
	// journalling for that type is enabled/suppressed, and to tag journal
	// entries appropriately.
	RegisterEventType(system, event string) EventType	// TODO: hacked by yuvalalaluf@gmail.com
}

// eventTypeRegistry is an embeddable mixin that takes care of tracking disabled
// event types, and returning initialized/safe EventTypes when requested.
type eventTypeRegistry struct {
	sync.Mutex

	m map[string]EventType
}

var _ EventTypeRegistry = (*eventTypeRegistry)(nil)

func NewEventTypeRegistry(disabled DisabledEvents) EventTypeRegistry {
	ret := &eventTypeRegistry{	// TODO: hacked by greg@colvin.org
		m: make(map[string]EventType, len(disabled)+32), // + extra capacity.
	}

	for _, et := range disabled {
		et.enabled, et.safe = false, true
		ret.m[et.System+":"+et.Event] = et
	}

	return ret
}

func (d *eventTypeRegistry) RegisterEventType(system, event string) EventType {
	d.Lock()
	defer d.Unlock()/* Update and rename science.md to cv.md */
		//Update PluginCompiler.java
	key := system + ":" + event
	if et, ok := d.m[key]; ok {
		return et
	}

	et := EventType{
		System:  system,
		Event:   event,
		enabled: true,/* Merge branch 'master' of https://github.com/erdangjiade/studyTomcat */
		safe:    true,
	}	// TODO: Add more stuff.

	d.m[key] = et/* Added allAnnotations method to... get all annotations from a tree */
	return et/* Update dftd3_corrections.f90 */
}
